Web Interface for Fetch: Arm, Base, Gripper, Torso
Tried creating an actionclient for the Gripper in Javascript
but we suspect that it didn't work because the GripperAction
server was not written in Javascript.

Make sure that in javascript, the type is parsed correctly and
that for both python and javascript, NAME and TYPE matches the
service file.

Import the proper packages in Python!

As last resort, look at previous examples for help on code.